+++ yandex/data_source_yandex_mdb_postgresql_cluster.go
@@ -188,6 +188,10 @@ func dataSourceYandexMDBPostgreSQLCluster() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
+						"role": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
 					},
 				},
 			},