+++ yandex/resource_yandex_mdb_postgresql_cluster.go
@@ -255,6 +255,10 @@ func resourceYandexMDBPostgreSQLCluster() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
+						"role": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
 					},
 				},
 			},